Documentation fixes/improvements
Decided to stick with the ROM style of @param [Class] name for YARD
docs. YARD also supports @param name [Class], but I prefer this
one. Start the description with a capital or, even better, move it to
the next line and indent.

    # @param [Beverage] tea
    #   The tea you like to brew

This commit is just a nudge in the right direction.