Make all Attributes based classes pp-able, make pp round-trip
Before we already had Resource#pp to get a nicely indented
representation of a Resource and its parts. By lifting this up to
Attributes it now becomes available in all Attributes based classes (all
data objects in Yaks).

The implementation of pp has also changed a little bit so the output can
be evaluated again to reconstruct the value. Great for testing and debugging!