BREAKING: drop symbol support for links in favor of lambda
Before instead of a link template we would allow a lambda (callable)
that returns a fully generated URI, but also a symbol, which would be
treated as a method name to be looked up like an object attribute, so
either as a method on the object or as a method on the mapper. (Or
something else if load_attribute was overridden).

This commit drops the latter option. Instead of

    link :foo, :bar

Use

    link :foo, -> { load_attribute(:bar) }

Although usually you would just use

    link :foo, -> { object.bar } # load directly from the object

or

    link :foo, -> { bar } # mapper has a bar method