Drop `form FormClass` in favor of `append_to(:forms, FormClass)`
Before and "create" style DSL method, like form or link, could either
take the arguments that would get passed to Form.create / Link.create,
or an actual Form/Link object. This is messy and led to issues.

Instead the mapper DSL is exposing append_to on the config object for
explicitly appending an object to the right list.