Make dynamic fields respect source order
In forms, field declarations are evaluated and stored at load time. This
is fine for most cases, but sometimes you need to be able to add form
fields during the mapping phase, e.g. when a `<select>` field's
available `<option>`s depend on the object being mapped.

For this we introduced "dynamic blocks". You can use these inside a form
declaration, they will be evaluated at map time, an receive the
to-be-mapped object as an argument. Inside you can make form field
declarations just like in the form itself.

    form :address do
      text :street

      dynamic do |customer|
        select :postage do
          customer.available_postage_options.each do |o|
            option o, value: o.id
          end
        end
      end
    end

The fields created by these dynamic blocks were being added at the end
of the form, instead of in the source order where they appear. This
meant you couldn't have a dynamic field inside a form unless you wrapped
the whole inside of the form in `dynamic`.

This commit fixes that, it also simplifies things in general by having
`Form::Config#fields` be a mixed array of `Field`, `Fieldset`, and
`DynamicField`, each responding to `to_resource_fields(mapper)`. This
can then simply be `flat_map`ped over.

Unified the ConfigBuilder for Form and Fieldset.