Introduce a separate Form::Legend class (both for Mapper and Resource)
Up to now `<legend>` form elements have been represented as Form::Field
instances, however, legends are much more contrained. They have a
content/label and that's it. Up to now we abused the :name attribute for
the label, but that's not accurate, the main reason is that that way all
Field DSL methods take the name as first argument.

Now legend takes a label, and optionally an `if`, and that's it.