Drop Yaks::FP, a remnant of a time when Yaks' code was even more obscure
The only method left was send_with_args, which at the time seemed a
great way to write less literal blocks. In reality even I had to look at
the implementation to replace it in the single call-site that was left,
and the literal form is infinitely more clear. Lesson learned.