Implement map_fields in terms of (a modified) fields_flat
This unifies the logic of recursively walking through form and fieldsets
to collect their fields. A side effect is that fields_flat returns an
enumerable instead of an array, which shouldn't be an issue in most
cases, although it is potentially a breaking change depending on how
calling code uses it.