Start using Ataru to test the README
Ataru is a tool for automated testing of your documentation. This is the
initial setup, and already some small fixes. We currently have 8 out of
30 code snippets in the README which give an error. Most likely because
they need some context that needs to be set up in ataru_setup.rb. I did
already find one mistake though, so it definitely works :)

Thanks @nerdbabe and @madziaf from the @CodePadawans for this awesome
tool <3