Get rid of the IANA registry CSV
While it some level nice to have, I no longer think this is the way to
go. First of all IANA is not complete, the microformats registry is
recommended. We were only using it in the HAL reader, for a check that
could be done much simpler.

We will want some better identifier handling in the future in yaks, but
this is not a good way to get there.