Make sure the acceptance tests actually work
It seems there were a couple of problems with the acceptance test, the
biggest being that the `let()` blocks defined in shared example groups
were interfering with each other. To prevent this I've wrapped them all
in `context {}` blocks.

Also added a new "list of quotes" fixture to test/document the behavior
of rendering a top level collection, and renamed the fixtures for
collection json to use the right naming convention.