Add specs for Yaks::Sinatra, classic and modular
These are simple specs: they verify that for a given Accept header, a
trivial Sinatra app with a trivial set of Yaks mappers will return a
viable response with the correct mime type and explicit charset. Both
a modular and a 'classic' Sinatra app are tested.

The MediaType class in `spec/integration_helper.rb` uses the HTTP spec's
definition of the Content-Type header (and other media type headers) to
parse a media type string and extract type and parameter parts. It's not
ludicrous-real-world-example proof, but it should be more than enough
for good-faith attempts at spec adherence of the kind expected from
Sinatra and from Yaks.

This commit also remodels the Yaks::Sinatra namespace so that it's
Yaks::Sinatra and not Sinatra::Yaks. Auto-registration for classic apps
is verified, as is explicit registration for modular apps.

Notes are added to the README regarding correct usage in both classic
and modular modes.

Files are Rubocop checked.